refactor: extract common E2E test patterns into harness helpers
Add push_screen(), type_into_text_input(), and dismiss_if_present()
helpers to reduce duplicated boilerplate across test phases. Remove
duplicate verify_receive_button_visible() call in phase_02 and
consolidate if/else branches in run_dpns_lookup().
